Refactor: Modularity for main(composable func)
