+++ app/src/main/java/com/codehanzoom/greenwalk/Main.kt
@@ -1,2 +1,34 @@
 package com.codehanzoom.greenwalk
 
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.navigation.compose.rememberNavController
+import com.codehanzoom.greenwalk.nav.BottomNavigation
+import com.codehanzoom.greenwalk.nav.NavigationGraph
+
+@Composable
+fun Main() {
+    val navController = rememberNavController()
+    Scaffold(
+        topBar = {
+            areaHeader()
+        },
+        bottomBar = {
+            BottomNavigation(navController = navController)
+        }
+    ) { innerPadding ->
+        Box(
+            modifier = Modifier
+                .padding(innerPadding)
+                .fillMaxWidth()
+                .fillMaxHeight()
+        ) {
+            NavigationGraph(navController = navController)
+        }
+    }
+}
